requirements.txt
Flask==1.0.2
Flask-Migrate==2.4.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
flask_marshmallow==0.10.0
marshmallow-sqlalchemy==0.16.2
psycopg2-binary==2.7.7
python-dotenv==0.10.1
celery==4.2.2
redis==3.2.1
apscheduler==3.6.0
requests==2.21.0
BeautifulSoup4==4.7.1